+++ src/features/post/inFeed/large/media/BlurOverlayMessage.tsx
@@ -2,36 +2,6 @@ import { css } from "@emotion/react";
 import styled from "@emotion/styled";
 import { IonIcon } from "@ionic/react";
 import { alertCircle } from "ionicons/icons";
-import { ReactNode } from "react";
-
-const BlurContainer = styled.div`
-  position: relative;
-`;
-
-const BlurContents = styled.div<{ blur: boolean }>`
-  ${({ blur }) =>
-    blur &&
-    css`
-      filter: blur(40px) brightness(0.8) saturate(1.2);
-
-      // https://graffino.com/til/CjT2jrcLHP-how-to-fix-filter-blur-performance-issue-in-safari
-      transform: translate3d(0, 0, 0);
-    `}
-`;
-
-interface BlurOverlayProps {
-  blur: boolean;
-  children: ReactNode;
-}
-
-export default function BlurOverlay({ blur, children }: BlurOverlayProps) {
-  return (
-    <BlurContainer>
-      <BlurContents blur={blur}>{children}</BlurContents>
-      {blur && <BlurOverlayMessage />}
-    </BlurContainer>
-  );
-}
 
 const MessageContainer = styled.div`
   // Safari bug where absolutely positioned content isn't viewable over
@@ -56,13 +26,13 @@ const MessageContainer = styled.div`
   container-type: size;
 `;
 
-const tallerImageQuery = "min-height: 150px";
-
 const WarningIcon = styled(IonIcon)`
   font-size: 42px;
+`;
 
-  @container (${tallerImageQuery}) {
-    margin-bottom: 8px;
+const showIfTaller = css`
+  @container (min-height: 150px) {
+    display: block;
   }
 `;
 
@@ -70,11 +40,11 @@ const Title = styled.div`
   font-size: 1.2em;
   font-weight: 600;
 
+  margin-top: 8px;
+
   display: none;
 
-  @container (${tallerImageQuery}) {
-    display: block;
-  }
+  ${showIfTaller}
 `;
 
 const Description = styled.div`
@@ -83,12 +53,10 @@ const Description = styled.div`
 
   display: none;
 
-  @container (${tallerImageQuery}) {
-    display: block;
-  }
+  ${showIfTaller}
 `;
 
-function BlurOverlayMessage() {
+export default function BlurOverlayMessage() {
   return (
     <MessageContainer>
       <WarningIcon icon={alertCircle} />
